requirements.txt
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0
joblib>=1.3.2
streamlit>=1.32.0
plotly>=5.19.0
requests>=2.31.0
python-dotenv>=1.0.0
deep-translator>=1.11.4
matplotlib>=3.8.0
seaborn>=0.13.2
